MSDK-441: consolidate form-encoded CharacterSet into shared extension
Address review: ATTNEventURLProvider already had a private
`formEncodedAllowedCharacters` (added in MSDK-354). Hoist it into a
shared `CharacterSet.attnFormEncodedAllowed` in Helpers/Extension so
both the legacy `/e` query builder and the v2 `/mobile` body encoder
use one definition instead of drifting copies.
